Interface rename and move PlotModifier and implementation to other package.
